Set allowedPeer with flag and set Viper defaults - cleanup and refactor code
