+++ pkg/pubsub/key_distribution.go
@@ -2,7 +2,10 @@ package pubsub
 
 import (
 	"encoding/json"
+	"errors"
 
+	libp2pCrypto "github.com/libp2p/go-libp2p/core/crypto"
+	masaCrypto "github.com/masa-finance/masa-oracle/pkg/crypto"
 	"github.com/sirupsen/logrus"
 )
 
@@ -11,22 +14,6 @@ const PublicKeyTopic = "public-key-topic"
 // PublicKeyMessage represents the structure of the public key messages.
 type PublicKeyMessage struct {
 	PublicKey string `json:"publicKey"`
-}
-
-// PublishPublicKey publishes the given public key to the public key topic.
-func (sm *Manager) PublishPublicKey(publicKey string) error {
-	msg := PublicKeyMessage{PublicKey: publicKey}
-	msgBytes, err := json.Marshal(msg)
-	if err != nil {
-		logrus.Errorf("Failed to marshal public key message: %v", err)
-		return err
-	}
-	if err := sm.gossipSub.Publish(PublicKeyTopic, msgBytes); err != nil {
-		logrus.Errorf("Failed to publish public key to topic %s: %v", PublicKeyTopic, err)
-		return err
-	}
-	logrus.Infof("Successfully published public key to topic %s", PublicKeyTopic)
-	return nil
 }
 
 // SubscribeToPublicKeyTopic subscribes to the public key topic to receive updates.
@@ -54,4 +41,23 @@ func (sm *Manager) SubscribeToPublicKeyTopic() error {
 		}
 	}()
 	return nil
+}
+
+// This method can be called directly after creating the Manager instance,
+// or at any other appropriate time when you need to publish or update the public key.
+func (sm *Manager) PublishNodePublicKey(publicKey string, data, signature []byte) error {
+	// Convert the publicKey string to a libp2p PubKey
+	pubKey, err := libp2pCrypto.UnmarshalPublicKey([]byte(publicKey)) // Use libp2pCrypto for UnmarshalPublicKey
+	if err != nil {
+		return errors.New("failed to unmarshal public key")
+	}
+
+	// Verify the signature using the VerifySignature function from your custom crypto package
+	isValid, err := masaCrypto.VerifySignature(pubKey, data, signature) // Use masaCrypto for VerifySignature
+	if err != nil || !isValid {
+		return errors.New("unauthorized: signature verification failed")
+	}
+
+	// Proceed to publish the public key if the signature is valid
+	return sm.PublishNodePublicKey(publicKey, data, signature)
 }
