+++ pkg/pubsub/manager.go
@@ -3,14 +3,11 @@ package pubsub
 import (
 	"bufio"
 	"context"
-	"errors"
 	"fmt"
 	"os"
 
-	libp2pCrypto "github.com/libp2p/go-libp2p-core/crypto"
 	pubsub "github.com/libp2p/go-libp2p-pubsub"
 	"github.com/libp2p/go-libp2p/core/host"
-	masaCrypto "github.com/masa-finance/masa-oracle/pkg/crypto"
 	"github.com/sirupsen/logrus"
 )
 
@@ -145,23 +142,4 @@ func StreamConsoleTo(ctx context.Context, topic *pubsub.Topic) {
 			logrus.Errorf("### Publish error: %s", err)
 		}
 	}
-}
-
-// This method can be called directly after creating the Manager instance,
-// or at any other appropriate time when you need to publish or update the public key.
-func (sm *Manager) PublishNodePublicKey(publicKey string, data, signature []byte) error {
-	// Convert the publicKey string to a libp2p PubKey
-	pubKey, err := libp2pCrypto.UnmarshalPublicKey([]byte(publicKey)) // Use libp2pCrypto for UnmarshalPublicKey
-	if err != nil {
-		return errors.New("failed to unmarshal public key")
-	}
-
-	// Verify the signature using the VerifySignature function from your custom crypto package
-	isValid, err := masaCrypto.VerifySignature(pubKey, data, signature) // Use masaCrypto for VerifySignature
-	if err != nil || !isValid {
-		return errors.New("unauthorized: signature verification failed")
-	}
-
-	// Proceed to publish the public key if the signature is valid
-	return sm.PublishPublicKey(publicKey)
 }
